Terminando testes automatizados de services

+++ src/app/services/DeleteArticleService.ts
@@ -1,4 +1,5 @@
 import { delay, inject, injectable, registry } from "tsyringe";
+import { AppError } from "../../errors/AppError";
 import { Article } from "../models/Article";
 import ArticlesRepository from "../repositories/ArticlesRepository";
 import { IArticlesRepository } from "../repositories/IArticlesRepository";
@@ -20,6 +21,13 @@ export class DeleteArticlesService {
 
     public async execute(id: number): Promise<void> {
         
+        const article = await this.repository.findById(id);
+
+        if(!article)
+        {
+            throw new AppError("Article not found");
+        }
+
         await this.repository.delete(id);
 
     }
